ci: change how release drafts are created in the workflow
Switched from manually creating a GitHub draft using `gh release` to using the `nekofar/create-github-release` action. This approach is clearer and more manageable. It allows you to set the release as a draft and conditionally marks it as a pre-release based on the presence of `-rc`, `-beta` or `-alpha` in the version.